+++ lib/core/router/bindings/password_reset_binding.dart
@@ -1,6 +1,6 @@
 import 'package:get/get.dart';
 import 'package:ondo/presentation/auth/password_reset/controllers/password_reset_completed_controller.dart';
-import 'package:ondo/presentation/auth/password_reset/controllers/password_reset_controller.dart';
+import 'package:ondo/presentation/auth/password_reset/controllers/password_reset_email_controller.dart';
 import 'package:ondo/presentation/auth/password_reset/controllers/password_reset_email_code_controller.dart';
 import 'package:ondo/presentation/auth/password_reset/controllers/password_reset_flow_controller.dart';
 import 'package:ondo/presentation/auth/password_reset/controllers/password_reset_password_controller.dart';
@@ -12,7 +12,7 @@ class PasswordResetBinding extends Bindings {
       () => PasswordResetFlowController(),
       fenix: true,
     );
-    Get.lazyPut<PasswordResetController>(() => PasswordResetController());
+    Get.lazyPut<PasswordResetEmailController>(() => PasswordResetEmailController());
     Get.lazyPut<PasswordResetEmailCodeController>(
       () => PasswordResetEmailCodeController(),
     );
